+++ light-desk/src/backend/components/group.ts
@@ -29,8 +29,16 @@ export class Group extends Component implements Parent {
   public addChild(component: Component) {
     this.children.push(component);
     component.setParent(this);
-    // TODO: allow children to have only one parent
-    // TODO: prevent loops
+    this.updateTree();
+  }
+
+  public removeChild(component: Component) {
+    const match = this.children.findIndex((c) => c === component);
+    if (match >= 0) {
+      const removed = this.children.splice(match, 1);
+      removed.map((c) => c.setParent(null));
+      this.updateTree();
+    }
   }
 
   public setTitle(title: string) {