+++ light-desk/src/frontend/util/react.ts
@@ -1,0 +1,3 @@
+export const calculateClass = (
+  ...args: (string | undefined | null | false)[]
+): string => args.filter((a) => !!a).join(' ');