+++ server/src/stage/plugins/virtual-output.ts
@@ -3,6 +3,8 @@ import * as ld from '@synesthesia-project/light-desk';
 
 import { Output, OutputContext, OutputKind, Plugin } from '.';
 
+const PIXEL_COUNT_MATCH = /^[0-9]+$/;
+
 const VIRTUAL_OUTPUT_CONFIG = t.type({
   pixels: t.number,
 });
@@ -10,15 +12,40 @@ const VIRTUAL_OUTPUT_CONFIG = t.type({
 type Config = t.TypeOf<typeof VIRTUAL_OUTPUT_CONFIG>;
 
 const createVirtualOutput = (
-  _context: OutputContext<Config>
+  context: OutputContext<Config>
 ): Output<Config> => {
   const group = new ld.Group();
-  const label = new ld.Label(`virtual!`);
+  const label = new ld.Label(`pixels:`);
   group.addChild(label);
+
+  const pixels = new ld.TextInput('');
+  group.addChild(pixels);
+
+  const update = new ld.Button('Update');
+  group.addChild(update);
+
+  update.addListener(async () => {
+    const value = pixels.getValue();
+    if (PIXEL_COUNT_MATCH.test(value)) {
+      context.saveConfig({
+        pixels: parseInt(value),
+      });
+    } else {
+      throw new Error(`Invalid pixel count: ${value}`);
+    }
+  });
+
+  let rects = new ld.Group();
 
   return {
     setConfig: (config) => {
-      label.setText(`virtual: ${config.pixels} pixel(s)`);
+      group.removeChild(rects);
+      rects = new ld.Group({ noBorder: true });
+      group.addChild(rects);
+      for (let i = 0; i < config.pixels && i < 10; i++) {
+        rects.addChild(new ld.Rect(ld.color.COLOR_RGB_WHITE));
+      }
+      pixels.setValue(`${config.pixels}`);
     },
     getLightDeskComponent: () => group,
   };