+++ light-desk/CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- `Group`: Allow components to be dynamically added
+  and removed after first render.
+- `Button`: Introduce error state that will trigger
+  when any listeners throw an error,
+  or return a promise that rejects.
+- `TextInput`: Introduce new generic text input component.
+
 ## [2.1.1] - 2022-01-11
 
 - Update ws and @types/ws